move tests to new file
