+++ packages/grafana-prometheus/src/components/metrics-browser/selectorBuilder.test.ts
@@ -1,19 +1,93 @@
 import { buildSelector, facetLabels } from './selectorBuilder';
 import { METRIC_LABEL, SelectableLabel } from './types';
 
-// Mock the dependencies
-jest.mock('../../language_utils', () => ({
-  escapeLabelValueInExactSelector: (value: string) => value,
-  escapeLabelValueInRegexSelector: (value: string) => value,
-}));
-
-jest.mock('../../utf8_support', () => ({
-  utf8Support: (value: string) => value,
-  isValidLegacyName: (name: string) => /^[a-zA-Z_:][a-zA-Z0-9_:]*$/.test(name),
-}));
-
 describe('selectorBuilder', () => {
   describe('buildSelector', () => {
+    it('returns an empty selector for no labels', () => {
+      expect(buildSelector([])).toEqual('{}');
+    });
+    it('returns an empty selector for selected labels with no values', () => {
+      const labels: SelectableLabel[] = [{ name: 'foo', selected: true }];
+      expect(buildSelector(labels)).toEqual('{}');
+    });
+    it('returns an empty selector for one selected label with no selected values', () => {
+      const labels: SelectableLabel[] = [{ name: 'foo', selected: true, values: [{ name: 'bar' }] }];
+      expect(buildSelector(labels)).toEqual('{}');
+    });
+    it('returns a simple selector from a selected label with a selected value', () => {
+      const labels: SelectableLabel[] = [{ name: 'foo', selected: true, values: [{ name: 'bar', selected: true }] }];
+      expect(buildSelector(labels)).toEqual('{foo="bar"}');
+    });
+    it('metric selector without labels', () => {
+      const labels: SelectableLabel[] = [
+        { name: '__name__', selected: true, values: [{ name: 'foo', selected: true }] },
+      ];
+      expect(buildSelector(labels)).toEqual('foo{}');
+    });
+    it('selector with multiple metrics', () => {
+      const labels: SelectableLabel[] = [
+        {
+          name: '__name__',
+          selected: true,
+          values: [
+            { name: 'foo', selected: true },
+            { name: 'bar', selected: true },
+          ],
+        },
+      ];
+      expect(buildSelector(labels)).toEqual('{__name__=~"foo|bar"}');
+    });
+    it('metric selector with labels', () => {
+      const labels: SelectableLabel[] = [
+        { name: '__name__', selected: true, values: [{ name: 'foo', selected: true }] },
+        { name: 'bar', selected: true, values: [{ name: 'baz', selected: true }] },
+      ];
+      expect(buildSelector(labels)).toEqual('foo{bar="baz"}');
+    });
+
+    describe('utf8 support', () => {
+      it('metric selector with utf8 metric', () => {
+        const labels: SelectableLabel[] = [
+          { name: '__name__', selected: true, values: [{ name: 'utf8.metric', selected: true }] },
+        ];
+        expect(buildSelector(labels)).toEqual('{"utf8.metric"}');
+      });
+
+      it('metric selector with utf8 labels', () => {
+        const labels: SelectableLabel[] = [
+          { name: '__name__', selected: true, values: [{ name: 'foo', selected: true }] },
+          { name: 'utf8.label', selected: true, values: [{ name: 'baz', selected: true }] },
+        ];
+        expect(buildSelector(labels)).toEqual('foo{"utf8.label"="baz"}');
+      });
+
+      it('metric selector with utf8 labels and metrics', () => {
+        const labels: SelectableLabel[] = [
+          { name: '__name__', selected: true, values: [{ name: 'utf8.metric', selected: true }] },
+          { name: 'utf8.label', selected: true, values: [{ name: 'baz', selected: true }] },
+        ];
+        expect(buildSelector(labels)).toEqual('{"utf8.metric","utf8.label"="baz"}');
+      });
+
+      it('metric selector with utf8 metric and with utf8/non-utf8 labels', () => {
+        const labels: SelectableLabel[] = [
+          { name: '__name__', selected: true, values: [{ name: 'utf8.metric', selected: true }] },
+          { name: 'utf8.label', selected: true, values: [{ name: 'uuu', selected: true }] },
+          { name: 'bar', selected: true, values: [{ name: 'baz', selected: true }] },
+        ];
+        expect(buildSelector(labels)).toEqual('{"utf8.metric","utf8.label"="uuu",bar="baz"}');
+      });
+
+      it('metric selector with non-utf8 metric with utf8/non-utf8 labels', () => {
+        const labels: SelectableLabel[] = [
+          { name: '__name__', selected: true, values: [{ name: 'foo', selected: true }] },
+          { name: 'utf8.label', selected: true, values: [{ name: 'uuu', selected: true }] },
+          { name: 'bar', selected: true, values: [{ name: 'baz', selected: true }] },
+        ];
+        expect(buildSelector(labels)).toEqual('foo{"utf8.label"="uuu",bar="baz"}');
+      });
+    });
+
     it('should return empty selector when no labels are selected', () => {
       const labels: SelectableLabel[] = [];
       expect(buildSelector(labels)).toBe('{}');
@@ -123,7 +197,9 @@ describe('selectorBuilder', () => {
           values: [{ name: 'localhost:9090', selected: true }],
         },
       ];
-      expect(buildSelector(labels)).toBe('http_requests_total{job=~"prometheus|node_exporter",instance="localhost:9090"}');
+      expect(buildSelector(labels)).toBe(
+        'http_requests_total{job=~"prometheus|node_exporter",instance="localhost:9090"}'
+      );
     });
 
     it('should handle non-legacy metric names with quotes', () => {
@@ -187,100 +263,46 @@ describe('selectorBuilder', () => {
     });
   });
 
-  describe('facetLabels', () => {
-    it('should update labels with possible values', () => {
-      const labels: SelectableLabel[] = [
-        { name: 'job', selected: true },
-        { name: 'instance', selected: false },
-      ];
-      
-      const possibleLabels: Record<string, string[]> = {
-        job: ['prometheus', 'node_exporter'],
-      };
-      
-      const result = facetLabels(labels, possibleLabels);
-      
-      expect(result).toHaveLength(2);
-      expect(result[0].name).toBe('job');
-      expect(result[0].values).toHaveLength(2);
-      expect(result[0].values?.[0].name).toBe('prometheus');
-      expect(result[0].values?.[1].name).toBe('node_exporter');
-      expect(result[0].hidden).toBe(false);
-      expect(result[0].facets).toBe(2);
-      
-      expect(result[1].name).toBe('instance');
-      expect(result[1].values).toBeUndefined();
-      expect(result[1].hidden).toBe(true);
-      expect(result[1].facets).toBe(0);
+  describe('facetLabels()', () => {
+    const possibleLabels = {
+      cluster: ['dev'],
+      namespace: ['alertmanager'],
+    };
+    const labels: SelectableLabel[] = [
+      { name: 'foo', selected: true, values: [{ name: 'bar' }] },
+      { name: 'cluster', values: [{ name: 'dev' }, { name: 'ops' }, { name: 'prod' }] },
+      { name: 'namespace', values: [{ name: 'alertmanager' }] },
+    ];
+
+    it('returns no labels given an empty label set', () => {
+      expect(facetLabels([], {})).toEqual([]);
     });
-    
-    it('should preserve selected values when facetting', () => {
-      const labels: SelectableLabel[] = [
-        { 
-          name: 'job', 
-          selected: true,
-          values: [
-            { name: 'prometheus', selected: true },
-            { name: 'grafana', selected: false },
-          ],
-        },
-        { name: 'instance', selected: false },
-      ];
-      
-      const possibleLabels: Record<string, string[]> = {
-        job: ['prometheus', 'node_exporter'],
-      };
-      
-      const result = facetLabels(labels, possibleLabels);
-      
-      expect(result[0].values).toHaveLength(2);
-      expect(result[0].values?.[0].name).toBe('prometheus');
-      expect(result[0].values?.[0].selected).toBe(true);
-      expect(result[0].values?.[1].name).toBe('node_exporter');
-      expect(result[0].values?.[1].selected).toBe(false);
-    });
-    
-    it('should keep existing values when facetting the same label', () => {
-      const labels: SelectableLabel[] = [
-        { 
-          name: 'job', 
-          selected: true,
-          values: [
-            { name: 'prometheus', selected: true },
-            { name: 'grafana', selected: false },
-          ],
-        },
-      ];
-      
-      const possibleLabels: Record<string, string[]> = {
-        job: ['prometheus', 'node_exporter'],
-      };
-      
-      const result = facetLabels(labels, possibleLabels, 'job');
-      
-      expect(result[0].values).toHaveLength(2);
-      expect(result[0].values?.[0].name).toBe('prometheus');
-      expect(result[0].values?.[1].name).toBe('grafana');
-    });
-    
-    it('should handle empty possible labels', () => {
-      const labels: SelectableLabel[] = [
-        { name: 'job', selected: true },
-        { name: 'instance', selected: false },
-      ];
-      
-      const possibleLabels: Record<string, string[]> = {};
-      
-      const result = facetLabels(labels, possibleLabels);
-      
-      expect(result).toHaveLength(2);
-      expect(result[0].hidden).toBe(true);
+
+    it('marks all labels as hidden when no labels are possible', () => {
+      const result = facetLabels(labels, {});
+      expect(result.length).toEqual(labels.length);
+      expect(result[0].hidden).toBeTruthy();
       expect(result[0].values).toBeUndefined();
-      expect(result[0].facets).toBe(0);
-      
-      expect(result[1].hidden).toBe(true);
-      expect(result[1].values).toBeUndefined();
-      expect(result[1].facets).toBe(0);
+    });
+
+    it('keeps values as facetted when they are possible', () => {
+      const result = facetLabels(labels, possibleLabels);
+      expect(result.length).toEqual(labels.length);
+      expect(result[0].hidden).toBeTruthy();
+      expect(result[0].values).toBeUndefined();
+      expect(result[1].hidden).toBeFalsy();
+      expect(result[1].values!.length).toBe(1);
+      expect(result[1].values![0].name).toBe('dev');
+    });
+
+    it('does not facet out label values that are currently being facetted', () => {
+      const result = facetLabels(labels, possibleLabels, 'cluster');
+      expect(result.length).toEqual(labels.length);
+      expect(result[0].hidden).toBeTruthy();
+      expect(result[1].hidden).toBeFalsy();
+      // 'cluster' is being facetted, should show all 3 options even though only 1 is possible
+      expect(result[1].values!.length).toBe(3);
+      expect(result[2].values!.length).toBe(1);
     });
   });
 });
